# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.",
        "title": "Government Template Platform V3.0 API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "Government Template Platform V3.0",
            "url": "https://github.com/snykk/go-rest-boilerplate"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Authorization header дахь JWT-ээс баталгаажуулагдсан хэрэглэгчийг уншиж, тохирох бичлэгийг буцаана (эхлээд in-memory кэш, олдоогүй үед Postgres).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Одоогийн хэрэглэгчийн профайлыг буцаах",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "User no longer exists",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "template_internal_http_datatransfers_responses.EIDCertificate": {
            "type": "object",
            "properties": {
                "issuer": {
                    "type": "string"
                },
                "key_type": {
                    "type": "string"
                },
                "not_after": {
                    "type": "string"
                },
                "not_before": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.EIDInfo": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDCertificate"
                },
                "civil_id": {
                    "type": "string"
                },
                "document_number": {
                    "type": "string"
                },
                "kyc_level": {
                    "description": "сертификатын түвшин",
                    "type": "string"
                },
                "national_id": {
                    "description": "регистрийн дугаар",
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.GoogleInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "linked_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eid": {
                    "description": "EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.\nНууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDInfo"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "first_name_en": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "full_name_en": {
                    "type": "string"
                },
                "google": {
                    "description": "Google нь холбогдсон Google account-аас хадгалсан профайл. Google\nхолбоогүй хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.GoogleInfo"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "last_name_en": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "/auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен (access token).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}